represent uid/gid as uint32, not uint16
this allows uid pools higher than 64k to work, as is often hit when
running Warden on each node of a test suite

e.g. uid pool start = 10000 * parallel-node

+++ linux_backend/src/wsh/pwd.h
@@ -8,8 +8,8 @@
 struct passwd {
   char *pw_name;   /* Username. */
   char *pw_passwd; /* Password. */
-  uint16_t pw_uid; /* User ID. */
-  uint16_t pw_gid; /* Group ID. */
+  uint32_t pw_uid; /* User ID. */
+  uint32_t pw_gid; /* Group ID. */
   char *pw_gecos;  /* Real name. */
   char *pw_dir;    /* Home directory. */
   char *pw_shell;  /* Shell program. */
